requirements_t.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
backports.entry-points-selectable==1.1.0
Brotli==1.0.9
CacheControl==0.12.6
cachy==0.3.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cleo==0.8.1
click==8.1.7
clikit==0.6.2
cmake==3.27.2
colorama==0.4.6
contourpy==1.1.0
crashtest==0.3.1
cryptography==41.0.3
cycler==0.11.0
distlib==0.3.2
exceptiongroup==1.1.3
filelock==3.0.12
fonttools==4.42.1
frozenlist==1.4.0
gevent==23.7.0
geventhttpclient==2.0.2
gitdb==4.0.10
GitPython==3.1.32
greenlet==2.0.2
grpcio==1.57.0
h11==0.14.0
html5lib==1.1
idna==3.4
importlib-resources==6.0.1
jeepney==0.8.0
Jinja2==3.1.2
keyring==21.8.0
kiwisolver==1.4.5
lit==16.0.6
lockfile==0.12.2
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
msgpack==1.0.2
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
nvidia-pytriton==0.2.5
opencv-python==4.8.0.76
packaging==20.9
pandas==2.0.3
pastel==0.2.1
pexpect==4.8.0
Pillow==10.0.0
pkginfo==1.7.1
platformdirs==2.3.0
poetry==1.1.8
poetry-core==1.0.4
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
py-cpuinfo==9.0.0
pycparser==2.21
pydantic==2.3.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
pylev==1.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-rapidjson==1.10
pytz==2023.3
pywin32-ctypes==0.2.0
PyYAML==6.0.1
pyzmq==23.2.1
requests==2.31.0
requests-toolbelt==0.9.1
scipy==1.10.1
seaborn==0.12.2
SecretStorage==3.3.3
sh==1.14.3
shellingham==1.4.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
tomlkit==0.7.2
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
triton==2.0.0
tritonclient==2.33.0
typing-inspect==0.6.0
typing_extensions==4.7.1
tzdata==2023.3
ultralytics==8.0.163
urllib3==2.0.4
virtualenv==20.7.2
webencodings==0.5.1
wrapt==1.15.0
yarl==1.9.2
zipp==3.16.2
zope.event==5.0
zope.interface==6.0
